Move Noop{Client,Server}Call from testing to api
To help with cross dependencies, we move these test classes to the test
fixtures of grpc-api.
