+++ services/src/test/java/io/grpc/protobuf/services/BinlogHelperTest.java
@@ -54,6 +54,8 @@ import io.grpc.Grpc;
 import io.grpc.Metadata;
 import io.grpc.MethodDescriptor;
 import io.grpc.MethodDescriptor.MethodType;
+import io.grpc.NoopClientCall;
+import io.grpc.NoopServerCall;
 import io.grpc.ServerCall;
 import io.grpc.ServerCallHandler;
 import io.grpc.Status;
@@ -67,8 +69,6 @@ import io.grpc.binarylog.v1.Message;
 import io.grpc.binarylog.v1.MetadataEntry;
 import io.grpc.binarylog.v1.ServerHeader;
 import io.grpc.binarylog.v1.Trailer;
-import io.grpc.internal.NoopClientCall;
-import io.grpc.internal.NoopServerCall;
 import io.grpc.protobuf.StatusProto;
 import io.grpc.protobuf.services.BinlogHelper.FactoryImpl;
 import io.grpc.protobuf.services.BinlogHelper.MaybeTruncated;
