+++ stub/build.gradle
@@ -19,7 +19,8 @@ dependencies {
         libraries.guava
     implementation libraries.errorprone.annotations
     testImplementation libraries.truth,
-            project(':grpc-testing')
+            project(':grpc-testing'),
+            testFixtures(project(':grpc-api'))
     signature libraries.signature.java
     signature libraries.signature.android
 }
